requirements.txt
# prod dependencies
gpiozero==2.0
RPi.GPIO==0.7.1
